+++ spec/untangled/client/impl/application_spec.cljs
@@ -167,26 +167,29 @@
     (app/sweep-merge {:a 1 :c 2} {:a 2 :c {:b 1}}) => {:a 2 :c {:b 1}}))
 
 (specification "Merge handler"
-  (let [triggers (atom {})
-        state (atom {:sa true})
-        rh (fn [env k v]
-             (assertions "return handler is passed state atom"
-               (-> env :state deref :sa) => true)
-             (swap! triggers assoc k v))]
+  (let [swept-state {:state 1}
+        data-response {:v 1}
+        mutation-response {'f {:x 1 :tempids {1 2}} 'g {:y 2}}
+        mutation-response-without-tempids (update mutation-response 'f dissoc :tempids)
+        response (merge data-response mutation-response)
+        rh (fn [state k v]
+             (assertions
+               "return handler is passed the swept state as a map"
+               state => swept-state
+               "tempids are stripped from return value before calling handler"
+               (:tempids v) => nil)
+             (vary-meta state assoc k v))]
     (when-mocking
       (app/sweep-merge t s) => (do
                                  (assertions
                                    "Passes source, cleaned of symbols, to sweep-merge"
-                                   s => {})
-                                 :return-of-sweep)
+                                   s => {:v 1})
+                                 swept-state)
 
-
-      (assertions
-        "Returns the result of an actual sweep-merge"
-        ;; Function under test:
-        (app/merge-handler state rh {} {'f {:x 1 :tempids {1 2}} 'g {:y 2}}) => :return-of-sweep
-
-        "triggers return-handler on symbols, passing in :result of return (while eliding tempids)"
-        (get @triggers 'f) => {:x 1}
-        (get @triggers 'g) => {:y 2}))))
+      (let [actual (app/merge-handler rh {} response)]
+        (assertions
+          "Returns the swept state reduced over the return handlers"
+          ;; Function under test:
+          actual => swept-state
+          (meta actual) => mutation-response-without-tempids)))))